Conflicts:
	TeamProject0503/src/net/member/action/MemberJoinAction.java
	TeamProject0503/src/net/member/action/MemberLoginAction.java

+++ TeamProject0503/src/net/vip/action/VipMovieDelete.java
@@ -2,6 +2,7 @@ package net.vip.action;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import net.vip.db.VipBean;
 import net.vip.db.VipDAO;
@@ -11,17 +12,21 @@ public class VipMovieDelete implements Action {
 
 	@Override
 	public ActionForward execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
-		System.out.println("GoodsDeleteAction execute");
-		request.setCharacterEncoding("UTF8");
-		
+		request.setCharacterEncoding("utf-8");
+		HttpSession session= request.getSession();
+		String id=(String)session.getAttribute("m_id");
+		if(id==null){
+			ActionForward forward= new ActionForward();
+			forward.setPath("./MemberLogin.me");
+			forward.setRedirect(true);
+			return forward;
+		}
 		VipDAO vipdao = new VipDAO();
 		VipBean vipbean = new VipBean();
 		
-		//int num=Integer.parseInt(request.getParameter("num"));
-		
 		vipdao.deleteVipMovie(Integer.parseInt(request.getParameter("v_num")));
 		
-		//이동
+		
 		ActionForward forward=new ActionForward();
 		forward.setRedirect(true);
 		forward.setPath("./VipMovieList.vi");
